+++ library/src/main/java/com/muellerwulff/simpleviewpager/Binding.kt
@@ -1,4 +1,4 @@
-package com.sihamark.simpleviewpager
+package com.muellerwulff.simpleviewpager
 
 import android.view.View
 import androidx.annotation.IdRes
@@ -12,14 +12,11 @@ fun setViewPager(tabLayout: TabLayout, @IdRes viewPagerId: Int) {
 	tabLayout.setupWithViewPager(viewPager)
 }
 
-private fun searchForViewPager(view: View, @IdRes viewPagerId: Int): ViewPager? {
-	return searchRecursively(view, viewPagerId, 0, 3)
-}
+private fun searchForViewPager(view: View, @IdRes viewPagerId: Int) =
+	searchRecursively(view, viewPagerId, 0, 3)
 
 private fun searchRecursively(
-	view: View, @IdRes viewPagerId: Int,
-	level: Int,
-	maxLevel: Int
+	view: View, @IdRes viewPagerId: Int, level: Int, maxLevel: Int
 ): ViewPager? {
 	if (level >= maxLevel) return null
 	val viewPager = view.findViewById<ViewPager>(viewPagerId)
@@ -27,10 +24,8 @@ private fun searchRecursively(
 		viewPager
 	} else {
 		val parent = view.parent as? View
-		if (parent != null) {
-			searchRecursively(parent, viewPagerId, level + 1, maxLevel)
-		} else {
-			null
+		parent?.let {
+			searchRecursively(it, viewPagerId, level + 1, maxLevel)
 		}
 	}
 }