+++ library/src/main/java/com/sihamark/simpleviewpager/Extensions.kt
@@ -1,6 +1,0 @@
-package com.sihamark.simpleviewpager
-
-import androidx.fragment.app.FragmentManager
-
-fun <T : Page> List<T>.toPagerAdapter(fragmentManager: FragmentManager) =
-	ListPagerAdapter(fragmentManager, this)